Deal with remote PTT type at the client end
Remote   PTT   must   always    be   either   RIG_PTT_RIG_MICDATA   or
RIG_PTT_NONE. Also take care not to  override any locally set PTT type
as it is feasible  to use a local hardware PTT at  the client end with
remote  CAT control.   Maybe an  odd arrangement  but Hamlib  does not
preclude it.  This is all  done while  preserving the accuracy  of the
ptt_type value in dump_state requests.