Updates to the TH-D7A/E back end
Allow for these rigs echo of successful set commands, no command
verification is requires nor necessary with TH style rigs.

Change rig_open to not use teh IF command with TH style rigs as it is
not supported and breaks communications for the next command.

Send the correct command to disable auto-information mode on TH style
rigs.

Ensure the correct command terminator is sent with all commands to TH
style rigs.

Fix an issue with MEM mode selection in the TH-D7.

Use the correct DCD detection command for the TH-D7.