Correct CWPitch set/get for Icom radios
Icom has two confusing ways to set/get CW Pitch. The first way is to read/write the knob setting using CI-V command 14 subcommand 9. The second way is just to set/get the actual value via mode command 10.

At some point the commands in hamlib changed from the former to the latter, but never adjusted the interpolation code of the former.

This commit fixes the set and get routines appropriately.